Fix transformer models in FP16

+++ fastai/text/models/transformer.py
@@ -15,7 +15,7 @@ class PositionalEncoding(nn.Module):
         self.register_buffer('freq', 1 / (10000 ** (torch.arange(0., d, 2.)/d)))
     
     def forward(self, pos:Tensor, bs:int=None):
-        inp = torch.ger(pos.float(), self.freq)
+        inp = torch.ger(pos, self.freq)
         enc = torch.cat([inp.sin(), inp.cos()], dim=-1)
         return enc
